Removed integer constraint from global message size

+++ edict--gui/src/main/java/guimodel/SystemSpecifications.java
@@ -22,7 +22,7 @@ public class SystemSpecifications {
 	
 	private int simulationDuration;
 	private String alias;
-	private int globalMessageSize;
+	private double globalMessageSize;
 	
 	public SystemSpecifications() {
 		super();
@@ -112,11 +112,11 @@ public class SystemSpecifications {
 		this.alias = alias;
 	}
 
-	public int getGlobalMessageSize() {
+	public double getGlobalMessageSize() {
 		return globalMessageSize;
 	}
 
-	public void setGlobalMessageSize(int globalMessageSize) {
+	public void setGlobalMessageSize(double globalMessageSize) {
 		this.globalMessageSize = globalMessageSize;
 	}
 	public boolean saveSystemSpecifications() {
